+++ node/src/accountant/test_utils.rs
@@ -14,8 +14,8 @@ pub fn make_receivable_account(n: u64, expected_delinquent: bool) -> ReceivableA
             n,
             if expected_delinquent { "d" } else { "n" }
         )),
-        balance: (n * 1_000_000_000) as i64,
-        last_received_timestamp: from_time_t(now - (n as i64)),
+        balance: (n * 1_000_000_000) as i128,
+        last_received_timestamp: from_time_t(now as i64 - (n as i64)),
     }
 }
 
@@ -23,8 +23,8 @@ pub fn make_payable_account(n: u64) -> PayableAccount {
     let now = to_time_t(SystemTime::now());
     PayableAccount {
         wallet: make_wallet(&format!("wallet{}", n)),
-        balance: (n * 1_000_000_000) as i64,
-        last_paid_timestamp: from_time_t(now - (n as i64)),
+        balance: (n * 1_000_000_000) as i128,
+        last_paid_timestamp: from_time_t(now as i64 - (n as i64)),
         pending_payment_transaction: None,
     }
 }
